+++ packages/frontend/src/components/Notes/Menu.vue
@@ -28,7 +28,7 @@
           density="compact"
           class="ml-2"
         />
-        <s-btn-icon v-if="props.exportUrl || props.performImport" size="small" density="compact">
+        <s-btn-icon v-if="props.exportUrl || props.performImport || (props.performDelete && props.selectedNotes)" size="small" density="compact">
           <v-icon icon="mdi-dots-vertical" />
           <v-menu activator="parent" eager :close-on-content-click="false" location="bottom right" class="context-menu">
             <v-list>
@@ -45,6 +45,21 @@
                 :export-url="props.exportUrl"
                 :name="props.exportName"
                 :disabled="props.readonly"
+              />
+              <btn-export
+                v-if="props.exportUrl && props.selectedNotes"
+                button-text="Export Selected"
+                :export-url="props.exportUrl"
+                :options="{notes: props.selectedNotes.map(n => n.id)}"
+                :name="props.exportName"
+                :disabled="props.readonly || props.selectedNotes.length === 0"
+              />
+              <btn-delete
+                v-if="props.performDelete && props.selectedNotes"
+                :delete="() => deleteNotes(props.selectedNotes!)"
+                :disabled="props.readonly || props.selectedNotes.length === 0"
+                button-variant="list-item"
+                :dialog-text="`Do you really want to delete ${props.selectedNotes.length} note(s)?`"
               />
             </v-list>
           </v-menu>
@@ -114,7 +129,9 @@ const props = defineProps<{
   createNote?: () => Promise<void>;
   exportUrl?: string;
   exportName?: string;
+  selectedNotes?: NoteBase[];
   performImport?: (file: File) => Promise<void>;
+  performDelete?: (note: NoteBase) => (Promise<void>|void);
 }>();
 
 const importBtnRef = useTemplateRef('importBtnRef');
@@ -128,6 +145,29 @@ function showSearch() {
 }
 function hideSearch() {
   search.value = null;
+}
+
+async function deleteNotes(notes?: NoteBase[]) {
+  if (!props.performDelete || !notes || notes.length === 0) {
+    return;
+  }
+
+  // Optimize delete: skip deleting children if parent is also being deleted
+  const deleteTasks = notes
+    .filter(n => !notes.some(n2 => n2.id === n.parent))
+    .map(n => (async () => {
+      try {
+        await Promise.resolve(props.performDelete!(n));
+      } catch (error: any) {
+        if (error?.status === 404) {
+          // Note was already deleted, ignore
+          return;
+        } else {
+          requestErrorToast({ error,  message: `Failed to delete note "${n.title}"` });
+        }
+      }
+    })());
+  await Promise.all(deleteTasks);
 }
 
 useKeyboardShortcut('ctrl+shift+f', () => showSearch());